+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/model/Labels.java
@@ -2,7 +2,7 @@
  * Copyright 2018, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.controller.cluster.resources;
+package io.strimzi.controller.cluster.model;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.strimzi.controller.cluster.operator.assembly.KafkaAssemblyOperator;
