+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/operator/assembly/AbstractAssemblyOperator.java
@@ -8,8 +8,8 @@ import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.strimzi.controller.cluster.operator.resource.ConfigMapOperator;
-import io.strimzi.controller.cluster.resources.AbstractCluster;
-import io.strimzi.controller.cluster.resources.Labels;
+import io.strimzi.controller.cluster.model.AbstractCluster;
+import io.strimzi.controller.cluster.model.Labels;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
