+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/operator/assembly/KafkaConnectAssemblyOperator.java
@@ -9,8 +9,8 @@ import io.fabric8.kubernetes.api.model.extensions.Deployment;
 import io.strimzi.controller.cluster.operator.resource.ConfigMapOperator;
 import io.strimzi.controller.cluster.operator.resource.DeploymentOperator;
 import io.strimzi.controller.cluster.operator.resource.ServiceOperator;
-import io.strimzi.controller.cluster.resources.KafkaConnectCluster;
-import io.strimzi.controller.cluster.resources.Labels;
+import io.strimzi.controller.cluster.model.KafkaConnectCluster;
+import io.strimzi.controller.cluster.model.Labels;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
