+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/operator/assembly/KafkaConnectS2IAssemblyOperator.java
@@ -12,8 +12,8 @@ import io.strimzi.controller.cluster.operator.resource.ConfigMapOperator;
 import io.strimzi.controller.cluster.operator.resource.DeploymentConfigOperator;
 import io.strimzi.controller.cluster.operator.resource.ImageStreamOperator;
 import io.strimzi.controller.cluster.operator.resource.ServiceOperator;
-import io.strimzi.controller.cluster.resources.KafkaConnectS2ICluster;
-import io.strimzi.controller.cluster.resources.Labels;
+import io.strimzi.controller.cluster.model.KafkaConnectS2ICluster;
+import io.strimzi.controller.cluster.model.Labels;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
