+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/operator/resource/AbstractResourceOperator.java
@@ -8,7 +8,7 @@ import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.KubernetesResourceList;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.fabric8.kubernetes.client.dsl.Resource;
-import io.strimzi.controller.cluster.resources.Labels;
+import io.strimzi.controller.cluster.model.Labels;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 import org.slf4j.Logger;
