+++ cluster-controller/src/test/java/io/strimzi/controller/cluster/operator/assembly/KafkaConnectAssemblyOperatorTest.java
@@ -11,8 +11,8 @@ import io.strimzi.controller.cluster.ResourceUtils;
 import io.strimzi.controller.cluster.operator.resource.ConfigMapOperator;
 import io.strimzi.controller.cluster.operator.resource.DeploymentOperator;
 import io.strimzi.controller.cluster.operator.resource.ServiceOperator;
-import io.strimzi.controller.cluster.resources.KafkaConnectCluster;
-import io.strimzi.controller.cluster.resources.Labels;
+import io.strimzi.controller.cluster.model.KafkaConnectCluster;
+import io.strimzi.controller.cluster.model.Labels;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
