+++ cluster-controller/src/test/java/io/strimzi/controller/cluster/operator/assembly/KafkaConnectS2IAssemblyOperatorTest.java
@@ -15,9 +15,9 @@ import io.strimzi.controller.cluster.operator.resource.ConfigMapOperator;
 import io.strimzi.controller.cluster.operator.resource.DeploymentConfigOperator;
 import io.strimzi.controller.cluster.operator.resource.ImageStreamOperator;
 import io.strimzi.controller.cluster.operator.resource.ServiceOperator;
-import io.strimzi.controller.cluster.resources.KafkaConnectCluster;
-import io.strimzi.controller.cluster.resources.KafkaConnectS2ICluster;
-import io.strimzi.controller.cluster.resources.Labels;
+import io.strimzi.controller.cluster.model.KafkaConnectCluster;
+import io.strimzi.controller.cluster.model.KafkaConnectS2ICluster;
+import io.strimzi.controller.cluster.model.Labels;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
